+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/DTOs/auth/AuthInfoDTO.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.data.repository.remote
+package com.realityexpander.tasky.data.repository.remote.DTOs.auth
 
 import com.realityexpander.tasky.common.AuthToken
 import com.realityexpander.tasky.common.UserId
