+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/IAuthApi.kt
@@ -1,10 +1,10 @@
 package com.realityexpander.tasky.data.repository.remote
 
-import com.realityexpander.tasky.BuildConfig
 import com.realityexpander.tasky.common.AuthToken
 import com.realityexpander.tasky.common.Email
 import com.realityexpander.tasky.common.Password
 import com.realityexpander.tasky.common.Username
+import com.realityexpander.tasky.data.repository.remote.DTOs.auth.AuthInfoDTO
 
 
 interface IAuthApi {
