+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/authApiImpls/AuthApiImpl.kt
@@ -1,8 +1,8 @@
 package com.realityexpander.tasky.data.repository.remote.authApiImpls
 
 import com.realityexpander.tasky.common.*
-import com.realityexpander.tasky.data.repository.remote.ApiCredentialsDTO
-import com.realityexpander.tasky.data.repository.remote.AuthInfoDTO
+import com.realityexpander.tasky.data.repository.remote.DTOs.auth.ApiCredentialsDTO
+import com.realityexpander.tasky.data.repository.remote.DTOs.auth.AuthInfoDTO
 import com.realityexpander.tasky.data.repository.remote.IAuthApi
 import retrofit2.HttpException
 import javax.inject.Inject
