+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/authApiImpls/TaskyApi.kt
@@ -1,9 +1,8 @@
 package com.realityexpander.tasky.data.repository.remote.authApiImpls
 
 import com.realityexpander.tasky.BuildConfig
-import com.realityexpander.tasky.common.AuthToken
-import com.realityexpander.tasky.data.repository.remote.ApiCredentialsDTO
-import com.realityexpander.tasky.data.repository.remote.AuthInfoDTO
+import com.realityexpander.tasky.data.repository.remote.DTOs.auth.ApiCredentialsDTO
+import com.realityexpander.tasky.data.repository.remote.DTOs.auth.AuthInfoDTO
 import com.realityexpander.tasky.data.repository.remote.IAuthApi
 import retrofit2.Response
 import retrofit2.http.*
