+++ app/src/main/java/com/realityexpander/tasky/presentation/agenda_screen/AgendaViewModel.kt
@@ -8,7 +8,7 @@ import com.realityexpander.tasky.R
 import com.realityexpander.tasky.common.Exceptions
 import com.realityexpander.tasky.di.AuthRepositoryFakeUsingProvides
 import com.realityexpander.tasky.di.AuthRepositoryProd_AuthApiProd_AuthDaoFake
-import com.realityexpander.tasky.presentation.util.UiText
+import com.realityexpander.tasky.presentation.common.util.UiText
 import com.realityexpander.tasky.domain.IAuthRepository
 import com.realityexpander.tasky.domain.validation.validateEmail.IValidateEmail
 import com.realityexpander.tasky.domain.validation.ValidatePassword
