+++ app/src/main/java/com/realityexpander/tasky/presentation/components/TextEntryField.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import com.realityexpander.tasky.R
 import com.realityexpander.tasky.presentation.ui.theme.textEntryFieldTextStyle
-import com.realityexpander.tasky.presentation.util.UiText
+import com.realityexpander.tasky.presentation.common.util.UiText
 
 @Composable
 fun TextEntryField(
