+++ app/src/main/java/com/realityexpander/tasky/presentation/login_screen/LoginScreen.kt
@@ -42,7 +42,7 @@ import com.realityexpander.tasky.presentation.components.PasswordField
 import com.realityexpander.tasky.presentation.destinations.AgendaScreenDestination
 import com.realityexpander.tasky.presentation.destinations.RegisterScreenDestination
 import com.realityexpander.tasky.presentation.ui.theme.TaskyTheme
-import com.realityexpander.tasky.presentation.util.UiText
+import com.realityexpander.tasky.presentation.common.util.UiText
 import kotlinx.coroutines.launch
 
 @Composable
